+++ examples/volumetric/tet_isos_slice.py
@@ -8,7 +8,7 @@ tmsh.add_scalarbar3d('chem_0  expression', c='k')
 
 vals = [0.2, 0.3, 0.8]
 tmsh.map_cells_to_points(["chem_0"])
-isos = tmsh.pointdata.select("chem_0").isosurface(vals)
+isos = tmsh.pointdata.select("chem_0").isosurface(vals).flat()
 
 slce = tmsh.slice(normal=(1,1,1)).lighting("off").lw(1)